Fix sv_cstr
sv_cstr was broken. The heuristic of plopping a '0' at the end or allocating
a new string which was then left to leak (because the caller had no idea
what happened behind the scenes so couldn't know what to free) is obvs
busted. So do the 1970s thing and let the caller provide a buffer to store
the cstr in.